Expecting a PHPUnit_Framework_Error with the mime_content_type(...) message instead of suppressing.
PHPUnit will convert notices/warnings/errors to exception by default, so
since we know the embedded resource doesn't exist, we can verify that
we'd see that message in a build situation and deployment would be
halted.

Also removed the try/catch, because I didn't see anywhere that an
InvalidArgumentException would be thrown. Even if it did get thrown, it
would probably be best to let the script blow up than to swallow it and
allow the build to complete broken.